pyproject.toml
[tool.poetry]
name = "cloudproject"
version = "0.1.0"
description = ""
authors = ["'starboyhassan' <'starboyhassan11@gmail.com'>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
pillow = "^10.0.0"
django-mptt = "^0.14.0"
pre-commit = "^3.3.3"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-phonenumber-field = "^7.1.0"
phonenumbers = "^8.13.18"
numpy = "^1.25.2"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
